+++ README.md
@@ -2,10 +2,10 @@
 
 A collection of helper packages to extend Docker Engine in Go
 
- Plugin type | Documentation | Description
- ------------|---------------|--------------------------------------------------
- Authz       | N/A                                                            | Extend API authorization mechanism
- Network     | [Link](https://docs.docker.com/engine/extend/plugins_network/) | Extend network management
- Volume      | [Link](https://docs.docker.com/engine/extend/plugins_volume/)  | Extend persistent storage
+ Plugin type   | Documentation | Description
+ --------------|---------------|--------------------------------------------------
+ Authorization | N/A                                                            | Extend API authorization mechanism
+ Network       | [Link](https://docs.docker.com/engine/extend/plugins_network/) | Extend network management
+ Volume        | [Link](https://docs.docker.com/engine/extend/plugins_volume/)  | Extend persistent storage
 
 See the [understand Docker plugins documentation section](https://docs.docker.com/engine/extend/plugins/).
