+++ authorization/README.md
@@ -1,20 +1,20 @@
-# Docker authz extension api.
+# Docker authorization extension api.
 
-Go handler to create external authz extensions for Docker.
+Go handler to create external authorization extensions for Docker.
 
 ## Usage
 
 This library is designed to be integrated in your program.
 
-1. Implement the `authz.Plugin` interface.
-2. Initialize a `authz.Handler` with your implementation.
-3. Call either `ServeTCP` or `ServeUnix` from the `authz.Handler`.
+1. Implement the `authorization.Plugin` interface.
+2. Initialize a `authorization.Handler` with your implementation.
+3. Call either `ServeTCP` or `ServeUnix` from the `authorization.Handler`.
 
 ### Example using TCP sockets:
 
 ```go
   p := MyAuthZPlugin{}
-  h := authz.NewHandler(p)
+  h := authorization.NewHandler(p)
   h.ServeTCP("test_plugin", ":8080")
 ```
 
@@ -22,7 +22,7 @@ This library is designed to be integrated in your program.
 
 ```go
   p := MyAuthZPlugin{}
-  h := authz.NewHandler(p)
+  h := authorization.NewHandler(p)
   h.ServeUnix("root", "test_plugin")
 ```
