+++ authorization/api_test.go
@@ -1,4 +1,4 @@
-package authz
+package authorization
 
 import "testing"
